Rename some model modules
Plural was used while singular was expected
