+++ zou/app/blueprints/crud/search_filters.py
@@ -1,4 +1,4 @@
-from zou.app.models.search_filters import SearchFilter
+from zou.app.models.search_filter import SearchFilter
 from zou.app.utils import permissions
 
 from .base import BaseModelResource, BaseModelsResource
