+++ zou/app/services/deletion_service.py
@@ -1,7 +1,7 @@
 from zou.app.models.comment import Comment
 from zou.app.models.entity import Entity
 from zou.app.models.subscription import Subscription
-from zou.app.models.notifications import Notification
+from zou.app.models.notification import Notification
 from zou.app.models.output_file import OutputFile
 from zou.app.models.preview_file import PreviewFile
 from zou.app.models.task import Task
